Improve performance

+++ WarriorsSnuggery/Game/Weapons/Warheads/ForceWarhead.cs
@@ -17,7 +17,7 @@ namespace WarriorsSnuggery.Objects.Weapons
 		[Desc("Range steps used for falloff.", "Defines at which range the falloff points are defined.")]
 		public readonly int[] RangeSteps = new[] { 0, 256, 512, 1024, 2048, 3096 };
 
-		readonly float maxRange;
+		readonly int maxRange;
 
 		public ForceWarhead(MiniTextNode[] nodes)
 		{
